+++ src/components/receivePayment/index.jsx
@@ -84,6 +84,7 @@ const ReceivePayment = ({
   vatPercent,
   disabled,
   valid,
+  invoiceInfo,
 }) => {
   const [sellerDropdown, openSellerDropdown] = useState(false);
   const { theme } = useTheme();
@@ -160,6 +161,7 @@ const ReceivePayment = ({
     transInfoArg,
     switchNetwork,
     switchAccount,
+    invoiceInfo,
   });
 
   const backend_API = new backendAPI();
@@ -386,7 +388,11 @@ const ReceivePayment = ({
   const selectInternalWallet = async () => {
     if (!Object.keys(user)?.length) {
       navigate("/login", {
-        state: { redirectUrl: `/pay/${transInfoArg.invoiceLink}` },
+        state: {
+          redirectUrl: transInfoArg.invoiceLink
+            ? `/pay/${transInfoArg.invoiceLink}`
+            : `/product/${transInfoArg.productLink}/pay`,
+        },
       });
     } else {
       await disconnect();